src/requirements.txt
astroid==2.11.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.2
dill==0.3.4
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-cloud-translate==3.7.2
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.45.0
grpcio-status==1.45.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.3.6
mccabe==0.7.0
pdfkit==1.0.0
pdfminer.six==20211012
pdfplumber==0.6.0
Pillow==9.0.1
platformdirs==2.5.1
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pylint==2.13.2
requests==2.27.1
rsa==4.8
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
Wand==0.6.7
wrapt==1.14.0
